third_party/blink/web_tests/platform/generic/fast/dom/Document/parent-node-interface-expected.txt

This tests that Document implements the ParentNode interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS doc.children.length is 0
PASS doc.childElementCount is 0
PASS doc.firstElementChild is null
PASS doc.lastElementChild is null
PASS doc.children.length is 0
PASS doc.childElementCount is 0
PASS doc.firstElementChild is null
PASS doc.lastElementChild is null
PASS doc.children.length is 1
PASS doc.childElementCount is 1
PASS doc.children[0] is b
PASS doc.firstElementChild is b
PASS doc.lastElementChild is b
PASS doc.children.length is 1
PASS doc.childElementCount is 1
PASS doc.children[0] is b
PASS doc.firstElementChild is b
PASS doc.lastElementChild is b
PASS successfullyParsed is true

TEST COMPLETE
